rename redux variable
