--- src/containers/Header/index.scss
@@ -1,43 +1,0 @@
-.header {
-    position: relative;
-    width: 100%;
-    height: 80px;
-    background: #dddddd;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-
-    .logo {
-        margin: 12px;
-        width: 50px;
-        height: 50px;
-        border-radius: 8px;
-        cursor: pointer;
-    }
-
-    .title {
-        font-family: sans-serif;
-        font-size: 24px;
-        font-weight: 600;
-        letter-spacing: 1px;
-    }
-
-    .bars {
-        margin: 12px;
-        padding: 8px;
-        width: 24px;
-        height: 24px;
-        background: #fff;
-        border-radius: 4px;
-        cursor: pointer;
-        transition: 0.2s;
-
-        path {
-            pointer-events: none;
-        }
-    }
-
-    .bars:hover {
-        opacity: 0.8;
-    }
-}
