--- src/redux/actions/header.js
@@ -1,4 +1,0 @@
-import { SHOWMENU, HIDEMENU } from "../constant"
-
-export const showMenu = () => ({ type: SHOWMENU })
-export const hideMenu = () => ({ type: HIDEMENU })
