+++ src/redux/actions/menu.js
@@ -1,0 +1,4 @@
+import { SHOW_MENU, HIDE_MENU } from "../constant"
+
+export const show_menu = () => ({ type: SHOW_MENU })
+export const hide_menu = () => ({ type: HIDE_MENU })
