+++ src/redux/store.js
@@ -1,5 +1,5 @@
 // import { configureStore } from "@reduxjs/toolkit"
 import { createStore } from "redux"
-import headerReducer from "./reducers/header"
+import headerReducer from "./reducers/menu"
 
 export default createStore(headerReducer)
